// src/types/user.d.ts
export interface User {
  id: string,
  email: string,
  avatar: string,
  username: string,
  firstName: string,
  lastName: string,
  password: string,
  birthdate: string,
}

export interface loggedUser {
  id: string,
  avatar: string,
  username: string,
  firstName: string,
  lastName: string,
  email: string,
  birthdate: string,
}